refactor(add) : 컴포넌트화 코드 분리, recoil 추가
