+++ package.json
@@ -10,14 +10,15 @@
   },
   "dependencies": {
     "axios": "^0.27.2",
-    "framer-motion": "^7.4.0",
     "formik": "^2.2.9",
+    "framer-motion": "^7.4.0",
     "next": "12.2.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.4.0",
     "react-query": "^3.39.2",
     "react-slick": "^0.29.0",
+    "recoil": "^0.7.6",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.5",
     "styled-reset": "^4.4.2",
