+++ src/components/ProductionAdd/index.js
@@ -1,0 +1,19 @@
+import ItemAdd from './ItemAdd';
+import ProductImageAdd from './ProductImageAdd';
+
+import ProductsAdd from './ProductsAdd';
+import StorePostAdd from './StorePostAdd';
+
+export default function ProductionAdd() {
+  const URL = '/api';
+
+  return (
+    <>
+      <h2>제품 등록</h2>
+      <ItemAdd />
+      <ProductImageAdd />
+      <StorePostAdd />
+      <ProductsAdd />
+    </>
+  );
+}
