+++ src/lib/data/productionAddAtoms.js
@@ -1,0 +1,21 @@
+import { atom } from 'recoil';
+
+export const storePostIdAtom = atom({
+  key: `storePostId`,
+  default: 0,
+});
+
+export const itemIdAtom = atom({
+  key: `itemId`,
+  default: 0,
+});
+
+export const contentImageIdAtom = atom({
+  key: `contentImageId`,
+  default: 0,
+});
+
+export const previewImageIdAtom = atom({
+  key: `previewImageId`,
+  default: 0,
+});
