+++ src/pages/_app.js
@@ -3,6 +3,7 @@ import '/styles/style.css';
 import Layout from 'components/layout';
 import { createGlobalStyle, ThemeProvider } from 'styled-components';
 import reset from 'styled-reset';
+import { RecoilRoot } from 'recoil';
 
 const GlobalStyle = createGlobalStyle`
 ${reset}
@@ -25,7 +26,9 @@ export default function MyApp({ Component, pageProps }) {
     <>
       <GlobalStyle />
       <ThemeProvider theme={theme}>
-        <Component {...pageProps} />
+        <RecoilRoot>
+          <Component {...pageProps} />
+        </RecoilRoot>
       </ThemeProvider>
     </>
   ) : (
@@ -33,7 +36,9 @@ export default function MyApp({ Component, pageProps }) {
       <GlobalStyle />
       <ThemeProvider theme={theme}>
         <Layout>
-          <Component {...pageProps} />
+          <RecoilRoot>
+            <Component {...pageProps} />
+          </RecoilRoot>
         </Layout>
       </ThemeProvider>
     </>
